--- works-api/DB.go
@@ -1,47 +1,0 @@
-package main
-
-import (
-	"database/sql"
-	"encoding/json"
-)
-
-func rowsToString(rows *sql.Rows) (string, error) {
-	columns, err := rows.Columns()
-	if err != nil {
-		return "", err
-	}
-
-	tableData := make([]map[string]interface{}, 0)
-
-	count := len(columns)
-	values := make([]interface{}, count)
-	scanArgs := make([]interface{}, count)
-	for i := range values {
-		scanArgs[i] = &values[i]
-	}
-
-	for rows.Next() {
-		err := rows.Scan(scanArgs...)
-		if err != nil {
-			return "", err
-		}
-
-		entry := make(map[string]interface{})
-		for i, col := range columns {
-			v := values[i]
-
-			b, ok := v.([]byte)
-			if ok {
-				entry[col] = string(b)
-			} else {
-				entry[col] = v
-			}
-		}
-
-		tableData = append(tableData, entry)
-	}
-
-	jsonData, _ := json.Marshal(tableData)
-
-	return string(jsonData), nil
-}
