+++ works-api/communicationGuacamole.go
