+++ works-api/properties.json
@@ -6,7 +6,7 @@
       "password":"Ablecloud1!"
     },
     "host":{
-      "address": "10.10.1.33",
+      "address": "10.10.1.24",
       "port": "3306",
       "protocol": "tcp"
     },
