+++ works-api/signature.go
@@ -17,8 +17,8 @@ func (s SortMoldParams) Len() int {
 	return len(s)
 }
 func (s SortMoldParams) Less(i, j int) bool {
-	for keyi, _ := range s[i] {
-		for keyj, _ := range s[j] {
+	for keyi := range s[i] {
+		for keyj := range s[j] {
 			return keyi < keyj
 		}
 	}
